\name{BF_output_networks}
\alias{BF_output_networks}
\title{ Generate DC networks }
\description{
  Generate differential co-expression modules and information of gene-gene interactions.
}
\usage{
BF_output_networks(
  dataExp,
  class,
  classlabel,
  BFobt,
  mst2file="MST2.txt",
  bfthr=6,
  corthres=0.3,
  echo=FALSE)
}
\arguments{
  \item{dataExp}{a data frame or matrix containing expression data. Columns correspond to genes and rows to samples. Such as in \code{\link{Compute_bf}}.}

  \item{class}{ a numeric or character vector contains the corresponding class information of dataExp. By far, package only accepts binary classes. }

  \item{classlabel}{ a numeric or character vector contains only two values. The first used to label class 1 and the second used to label class 2. the order of these two elements must be kept the same throughout the analyses. }

  \item{BFobt}{ a \code{BFobt} object generated by \code{\link{BF_WGCNA}}.}
 
  \item{mst2file}{ specify a file to output the generated union of First and Second Minimal Spanning tree (MST2). The default file name is  "MST2.txt".}

  \item{bfthr}{ Bayes factor threshold. Only gene pairs with Bayes factor higher than \code{bfthr} threshold can be considered. The default value is 6. }


  \item{corthres}{ Pearson Correlation Coefficient (PCC) threshold. It's used to classify edges into up, down and unchange classes. Edges with PCC between -corthres to +corthres is labeled as unchange, edges with PCC higher than corthres is labeled as up, and  edges with PCC smaller than corthres is labeled as down.}

  \item{echo}{ a logical variable, indicates whether print processing information or not. If \code{TRUE}, it will print processing information for every 1000 gene pairs. }

}
\details{
  The function returns a list containing (1) information for DC gene nodes, including module assignment and weight assignment.
                                         (2) information for DC gene pairs, including a full network built based on DC modules.  
  And the essential links among genes represented by a union of First and Second Minimal Spanning Tree (MST2) will be outputed into a file specified by argument mst2file. 
}
\value{
 A list containing \code{genegroups} and \code{network}. 
 \item{genegroups}{ containing following 4 columns:}
                   \enumerate{
                    \item gene: A character vector giving gene ids. 
                    \item colors: A character vector giving color labels of DC modules. 
                    \item labels: A numerical vector giving labels of DC modules, start from 0.
                    \item gene_weight: A numerical vector giving weight for each gene. 
 }
 \item{network}{ containing following 16 columns:}
                   \enumerate{
                    \item geneid1: A character vector giving ids for gene1.  
                    \item geneid2: A character vector giving ids for gene2.  
                    \item groupid: A numerical vector giving labels of DC modules for gene1 and gene2, connected genes belong to same DC modules.
                    \item type: A character vector indicating which model is selected for gene1 and gene2.
                    \item gene1_weight: A numerical vector giving weight for gene1.
                    \item gene2_weight: A numerical vector giving weight for gene2.
                    \item bf.value: A numerical vector giving pair-wise Bayes factors for gene1 and gene2.
                    \item pc1: A numerical vector giving Pearson Correlation Coefficient 
                    of gene1 and gene2 under class 1.
                    \item pc2: A numerical vector giving Pearson Correlation Coefficient 
                    of gene1 and gene2 under class 2.
                    \item cortypes: A character vector indicating the direction of correlation changing in terms of Pearson Correlation Coefficient. Compare the difference of \code{pc1} and \code{pc2} with \code{corthres}, it gives values from ("down","up","unchange").
                    \item mean_g1_c1:  A numerical vector giving the mean expression for gene1 under class 1.
                    \item mean_g1_c2:  A numerical vector giving the mean expression for gene1 under class 2.
                    \item mean_foldchange_gene1:  A numerical vector giving the fold change of mean expression for gene1 between class 1 and class 2.
                    \item mean_g2_c1:  A numerical vector giving the mean expression for gene2 under class 1.
                    \item mean_g2_c2:  A numerical vector giving the mean expression for gene2 under class 2.
                    \item mean_foldchange_gene2:  A numerical vector giving the fold change of mean expression for gene2 between class 1 and class 2.
                    }
}

\keyword{ package }

