\name{BF_similarity}
\alias{BF_similarity}
\title{ Convert pair-wise Bayes factors into network adjacency}
\description{
 Convert pair-wise Bayes factors into network adjacency (values between 0 and 1) which can be used as an input for other methods.
}
\usage{
BF_similarity(bfmatrix,softPower,bfthr=6,keepedges)
}
\arguments{
  \item{bfmatrix}{ a data frame containing information of pair-wise bayes factors generated by \code{\link{Compute_bf}}. Refer to \link{Compute_bf} for details.}
  
  \item{softPower}{ soft thresholding power. Such as the argument \code{power} in package WGCNA}

  \item{bfthr}{ Bayes factor threshold. The default value is 6. }
  
  \item{keepedges}{ a numerical variable indicating how many top edges to be remained in terms of Bayes factor. \code{keepedges} is invalid when is set as 0. When \code{keepedges} is not 0, it is prior to \code{bfthr}. The default value is the number of genes in \code{dataExp}. }  
}
\details{

  The function generates an adjacency matrix indicating the pair-wise differential co-expression under two different classes measured by pair-wise Bayes factors. It transfers pair-wise Bayes factor matrix \code{bfmatrix} into an adjacency matrix which can be used as an input for other methods. The arguments \code{bfthr} and \code{keepedges} are considered as hard threshold. When the hard threshold is set, the argument \code{softPower} for soft thresholding is usually set as 1 and only gene pairs which satisfy the bfthr threshold or keepedges threshold are considered in the clustering procedure. 

}
\value{

An adjacency matrix. If \code{bfmatrix} contains \var{n} different genes,  the adjacency matrix is of dimensions \var{n} times \var{n}.

}
\keyword{ package }

