\name{sigDCpair_SFS}
\alias{sigDCpair_SFS}
\title{The second step of selecting significant DC gene pairs. }
\description{
 It's for the second step of selecting significant DC gene pairs, after this procedure the final significant DC gene pairs are generated.
}
\usage{
sigDCpair_SFS(
  dataExp,
  class,
  classlabel,
  sigDC,
  DC_acc="DC_pair_acc.txt",
  bfthr=6,
  by=50,
  LOOCV=TRUE,
  valid=NULL,
  valid_class=NULL
  )
}
\arguments{

  \item{dataExp}{ a data frame or matrix containing expression data. Columns correspond to genes and rows to samples. Such as in \code{\link{Compute_bf}}.}

  \item{class}{ a numeric or character vector contains the corresponding class information of dataExp. By far, package only accepts binary classes. }

  \item{classlabel}{ a numeric or character vector contains only two values. The first used to label class 1 and the second used to label class 2. the order of these two elements must be kept the same throughout the analyses. }

  \item{sigDC}{ a data frame contains the selected gene pairs generated from \code{\link{sigDCpair_st1}}. }  
  
  \item{DC_acc}{ specify a file to output the results. The default fila name is "DC_pair_acc.txt". Columns in the file are splited by 'Tab'.}  

  \item{bfthr}{ Bayes factor threshold. Only gene pairs with Bayes factor higher than \code{bfthr} are considered. The default value is 6.}
 
  \item{by}{ an integer indicates the increment of gene pairs. The default value is 50.}

  \item{LOOCV}{logical: If \code{TRUE}, leave-one-out cross-validation (LOOCV) is used to monitor the accuracy of top N subset of gene pairs, otherwise a validation data must be supplied in argument \code{valid} and the accuracy is monitored on the validation data. The default value is \code{TRUE}.}

  \item{valid}{'NULL' (default) or data frame or matrix containing expression of validation data. Columns correspond to genes and rows to samples. Only used when \code{LOOCV} is \code{FALSE}. Must have compatible columns (order of columns and dimensions of columns) with \code{dataExp}. }
  
  \item{valid_class}{ a character vector containing class information for validition data. It is sorted according to the rows of \code{valid}.}


}
\details{
  The function generates a file specified in argument \code{DC_acc} which can be used to decide the final significant DC gene pairs.
  In this step, a generalized sequential forward selection (SFS) algorithm [1] is used on the candidate significant DC gene pairs generated by \code{\link{sigDCpair_st1}}. The gene pairs in \code{sigDC} are sorted in descending order of their corresponding scores. Start with an initial subset of gene pairs with size N=1, the next \code{by} gene pairs are repeatedly added to the subset of gene pairs, N=N+\code{by}. The goodness of top N subset of gene pairs are characterized by the leave-one-out cross-validation (LOOCV) or a validation dataset to monitor the accuracy for the top N gene pairs, and the accuracy is calculated by DC-based prediction model. The output file consist of two columns, the first indicates the number of top gene pairs and the second indicates the accuracy. According to this file, an accuracy curve can be plotted, than the decision of number N can be made as the final number of selected DC gene pairs. The final significant DC gene pairs is the top N gene pairs in \code{sigDC}.

}
\note{The running time of this function is closely related to the total number of candidate gene pairs generated from \code{\link{sigDCpair_st1}} and the argument \code{by}.}
\references{ 
[1] Whitney AW. A direct method of nonparametric measurement selection. Computers, IEEE Transactions on. 1971;100:1100-3.
}
\keyword{ package }

