\name{sigDCpair_st1}
\alias{sigDCpair_st1}
\title{ First step of selecting significant DC gene pairs. }
\description{
 Generate the candidate gene pairs selected from the first step of selecting significant DC gene pairs.
}
\usage{
sigDCpair_st1(
  BFobt,
  mst2="mst2",
  bfthr=6,
  sparse=6,
  weight_cutoff=0.8
  )
}
\arguments{
  \item{BFobt}{ a \code{BFobt} object generated by \code{\link{BF_WGCNA}}.}

  \item{mst2}{ an input file specified to provide the MST2 skeletons of the DC structures. It must be the resulting 
               MST2 file of \code{\link{BF_output_networks}}.}

  \item{bfthr}{ Bayes factor threshold. Only gene pairs with Bayes factor higher than \code{bfthr} are considered. The default value is 6.}
  
  \item{sparse}{ a tuning parameter controls the end-node weights in order to make a sparser subsets of gene pairs passing the \code{weight_cutoff} threshold. The defualt value is 6.}
  \item{weight_cutoff}{ threshold of end-node weights. The default value is 0.8. }

}
\details{
  The function generates the candidate gene pairs selected from the first step of selecting significant DC gene pairs. The result is a data frame contains three columns, the first is a character vector representing ids for gene1, the second is a character vector representing ids for gene2 and the last is a numeric vector representing assignment scores for each gene pair. The result is in a descending order according to the score. 
  The score is caculated by the geometric mean of the two end-node weights multiplied by the Bayes factor corresponding to the gene pair.
  In this procedure, each edges of MST2 stored in the file specified by argument mst2 will be searched, and those that satisfy the two thresholds, \code{bfthr} and \code{weight_cutoff}, will be remained.
  Here, the tuning parameter \code{sparse} and \code{weight_cutoff} are crucial to select the gene pairs. A higher \code{sparse} and higher \code{weight_cutoff} will result in a smaller set of 
  gene pairs with higher prediction power. 
}
\value{

   a data frame contains the candidate gene pairs from the first step of selecting significant DC gene pairs. It contains 3 columns: \enumerate{\item geneid1: A character vector representing ids for gene1.
   \item genedi2: A character vector representing ids for gene2. \item score:  A numeric vector representing assignment scores for each gene pair.} The data frame is sorted in a descending order according to the score.  
}
\references{ 
Rahmatallah Y, Emmert-Streib F, Glazko G. Gene Sets Net Correlations Analysis (GSNCA): a multivariate differential coexpression test for gene sets. Bioinformatics. 2014;30:360-8.
}
\keyword{ package }

