\name{SimulationSmall}
\alias{SimulationSmall}
\docType{data}
\title{small simulation data for test.}
\description{
This matrix gives class information and expression data generated by a simulation procedure.
}
\usage{data(SimulationSmall)}
\format{

A 200 x 241 matrix.  The first column (class) lists class labels, 1 indicate class 1 and 2 indicate class 2. Other columns correspond to genes. The rows are correspond to samples. There are 7 co-expression modules in total, each contains 20 genes, and 100 randomly generated genes. Module1-module2 have the same distribution in both class 1 and class 2, so they are stable modules. Module3 and module4 represent the intra-module DC, genes in these modules have differential co-expression patterns under different classes. Genes in module3 only correlated under one class. Genes in module4 highly correlated under one class, while less correlated under the other class. Module5 represents the shift scenario of DC, in which the correlations of genes under different classes do not change while their expression levels have an overall increase or decrease. Module6 and module7 represent the inter-module DC. Genes within each module have conserved co-expression patterns under two classes, but have a cross scenario between module6 and module7.   
}
\examples{
data(SimulationSmall)
head(SimulationSmall)
}
\keyword{datasets}

